+++ CHANGELOG.md
@@ -19,5 +19,6 @@ v1.0.1
 - Minor correction
 - Mixed elements for search items
 - Shortcuts methods on helpers classes
-
-    
+- Added __repr__ for lists
+- fixed relations
+- Detail client may return a list of element when retrievind multiple values for identifier